Add MultiplexCompletionSource with a WeightedRandomSelector
Refs #3

This lets us control which sources get which data, and then we
can choose between separate sources of completions. Right now we
just have a basic random, weighted selector, but in the future we
could use something like markov chains to get some real probability.

Right now we use some made-up heuristics where we weight user-input
tokens more highly for the first word and, to a lesser extent, the
second word. After that we just split it evenly.