Refactor TestableJudoCore so we don't have to edit it constantly
It now delegates to a proxy so we can just override the methods
we actually care about overriding, instead of constantly having
to add stubs whenever we add a method to IJudoCore.

Speaking of which, IJudoCore (and JudoCore, which implements it)
is becoming a bit god-like. Might be a good candidate for
refactoring at some point....