Refactor RecencyCompletionSource to use LinkedHashMap
The old implementation used a TreeMap with a container object, so
storage complexity was ~2N; The LinkedHashMap stores nodes in an
array, so it's also ~2N. However, insertions/etc go from O(log(N))
to O(1). Plus, the TreeMap implementation didn't de-duplicate keys,
so there was a lot of wasted space---this is especially problematic
for our use case, since we're highly likely to have *many* duplicate
keys. LinkedHashMap *does* de-dup correctly, so it can store many
more unique keys.

The problem is that, by default, iteration happens with
least-recently-used-first. This is great when you want to evict LRU
elements for a cache, but if you actually want to iterate over the
*most*-recently-used, as we do... it's a problem. We can't put a
subclass in the java.util package to provide a reverse iterator, due to
SecurityManager restrictions thereof, but we *can* use reflection to
access the `tail` field and iterate that way. Since iteration of
completion suggestions is not a hot code path, this overhead is minimal
and seems worth it---much better than accepting the memory overhead of a
parallel array that we periodically copy elements to and read in reverse
order.