Support tab completion in output search mode
Closes #23

It's not technically traditional in vim, but I think it makes sense.
We ONLY tab-complete from received output, which I think makes sense;
it's conceivable, but unlikely you'd want to search for your own
commands. If that comes up, we should probably not use the same
completion source used for input, but rather one that weights
output slightly higher