Fix a few bugs with suggestion handling
In particular, we didn't reset suggestions when entering a mode,
so exiting insert mode and clearing the input field via dd or
similar could cause confusion if you didn't type anything before
attempting tab complete. A test was added to cover this (although
not *every* mode) and the fix was applied to all modes with tab
complete.