Refactor and fix mccp inflation on some servers
Lusternia, for example, flushes compressed data *immediately* after
the MCCP enable sequence, so the beginning of the compressed sequence
was in the same buffer and thus not decompressed. Furthermore, it
doesn't always "finish" the compression gracefully when disconnecting
the client, so we have to do some manual reading to handle that
ourselves without throwing errors.