Add preliminary support for registers (#39)
* Add basic support for the unnamed register

Also prepare for supporting clipboard registers, and using them
by default if set to do so.

* Support accessing other registers; improve clipboard registers

* Implement the black hole register

This one is simple enough. We can implement the numbered registers
later

* Lazy-init the clipboard ref

This is mostly an attempt to fix Travis tests, but seems reasonable
enough

* Add the clipboard=unnamed setting

It turns out it's just the first access to the clipboard that's slow;
subsequent uses are nice and snappy. Included are some attempts at
ameliorating this initial access slowdown, but I'm not sure if we
can completely hide *all* the gross.