Preliminary mapping support (#40)
* Add preliminary world-mapping (NOT keymapping, of course) classes

Refs #20

Includes an early MapGrid for use visualizing the map, and a MapFormat
that can read and write TinTin++ map files.

Next we'll need something that visualizes using the MapGrid, and
finally expose map creation to scripting. We'll probably also
want to try a simple automap() command if the server supports MSDP
with room vnum and exits.

* Add simple map renderer

Refs #20

* Add MapManager to hold onto (and load, etc.) the current Map

* Expose the IJudoCore obj via interface; support deleting map rooms

Refs #20

By exposing IJudoCore (and others) in this way, we get a much
simpler way to expose scripting methods. This is probably the only
way we'll expose mapping, since there are potentially just so many
methods.

* Add primitive `map:autoroom` feature; prep for `map:automagic`

map:automagic will attempt to use MSDP to automagically create
and follow rooms. If/when we support GMCP, we could use that, too.

* Add some early attempts at auto-mapping

Not sure if there are any MUDs that actually do it the way the
MSDP spec suggests (and which is the only way we handle it right now)
but... it's a start, I guess.

* Fix test

* Clear the events manager on disconnect

* Support rendering maps; add early DirectionalMapStrategy impl.

Refs #20

* Attempt to improve DirectionalMapStrategy

* Revert accidental changes to README

* Support optionally providing a target window for map rendering