Undo/Redo and `.` Repeat (#42)
* Begin very rough sketch of basic delete undo functionality

Can't undo insert yet, but should be doable. A "better" implementation
would probably be to have a Command interface that knows how to
apply and undo itself, but this way is simpler to implement, I think;
we would have to do a whole Command composition thing; it'd be a mess.
Most of our changes break down to deletes and inserts anyway....

* Add preliminary support for undoing insertions and replacements

* Add super preliminary, experimental "repeat change" support

* Begin support for `.` (repeat) action

The `f` motion doesn't record it's target properly yet, so
`cfd` changes don't repeat properly.

Also, it seems that while bash supports `u`, it only supports
`.` to repeat, no redo. I think I'd still like to support it
somehow, but it may not be the end of the world if we don't right
away.

* Support repeating find motions

* Support "redo"

* Change default mapping for rsearch mode to <ctrl s>

For rSearch or hiStory. It's a bit meh, but it feels better for
the default to match Vim, and nyway users can remap `<ctrl s>`
to `<ctrl r>` (or whatever) if they like.

* Prevent stackoverflow with undo-redo-undo-redo

* Don't include motions, etc. in changes

* Include counts and register use in changes

* Update VimCompat with recent changes