Refactor rendering to pull JLine dependency out of Core (#55)
* Begin experimenting with refactored UI models

Refs #51

* Continue refactor with proper Connection Factory

* Begin work on JLine render module

* Begin implementing and testing FlavorableStringBuilder `append`

* Continue testing and fixing FlavorableStringBuilder

* Add another edge case test

* Begin testing JLineDisplay

* Prevent accidentally writing out of Display bounds

* More bounds checking

* Begin testing JLineWindow's rendering

* Fix blank line handling in JLineWindow rendering

* Add more JLineWindow rendering Tests

* Add BaseJudoWindow2; prepare to support scrolling in JLineWindow

* Begin work on new word wrap handling to support scrolling in JLineWindow

* Fix up (and clean up) word wrap algorithm

* Implement basic scrolling in JLineWindow

* Add initial support for scrolling to a specific buffer position

* Implement basic in-buffer search handling

* Import initial Tabpage implementation

* Begin importing JLineRenderer implementation to JLineRenderer2

* Fix TerminalInputLineHelper with "preserveWhitespace" splitting option

* Begin initial implementation of top-level rendering + tests

Status line doesn't render properly yet, and we need to verify the
finaly cursor placement....

* Fix status window rendering

* Add method for testing cursor positioning

* Begin implementation of JudoConnection2 using coroutines

* Add preliminary ANSI -> FlavorableCharSequence parsing

* Fix FlavorableStringBuilder.plusAssign

* Improve newline normalization

* Implement "Trailing Ansi" test

* Parse more ANSI colors

* Trigger a render when a change is made to a buffer or Window

For symmetry with scripting

* Render cursor position into JLineDisplay to simplify interactions

* Make JudoRenderer2 responsible for allocating buffers

* Add explicit test for scrolling with content

* Use the JudoRenderer2 interface instead of JLineRenderer2 directly

Fixes mock{} usage in tests

* Fill out `search with no result` unit test

* Fix status line rendering + import some tests

* Update semantics of JudoBuffer2 to expect a \n at the end of each line

Simplifies append from the network semantics

* First pass swapping in new render model

Most tests pass but a few are ignored for now. Priorities:

1. Update remaining tests for new render model
2. Implement readKey()
3. Implement CommonsNetConnection2, possibly pulling things up into
    BaseJudoConnection2 if possible
4. Test!!
5. Clean out old stuff and remove any remaining dependencies on JLine
    from `core` module
6. Rename all relevant model interfaces to drop the `2` suffix
    (probably?)
7. Merge!!

* Begin support for newlines in append()

The assumption that we would *only* pass single lines to Buffer.append()
was convenient but not good: we would have to add special handling to
each scripting implementation, which seems dumb. This handles many
cases, but one of the JudoCore tests is still ignored for handling the
case where the previous line in the buffer was incomplete before the
split.

* Improve handling of line splitting to fix some pending tests

* Fix some disabled tests; improve resize handling

* Resurrect a possibly-useful old JLineRendererTest; clean up others

* Move CommonsNetConnection to own project; drop dependency from core

Also, update it for JudoConnection2, of course

* Drop old JudoConnection; replace with (renamed) JudoConnection2

* Fix default connect timeout

* Refactor our "simple" MapRenderer to be based on IJudoBuffer2

Now it can be used with any renderer!

* Import readKey() from JLineRenderer

But we house it in its own class, now

* Change to use FlavorableCharSequence everywhere possible

Refs #51

Includes a new `FlavorableCharSequence.toAnsi()` util, because we still
need to be able to pass an ANSI string to prompt handlers, but in
general this cleans up a bunch of things that had to know about ANSI
that shouldn't need to.

This is probably the core change of this branch that generally makes
ANSI knowledge unnecessary in JudoCore, since we can just use
FlavorableCharSequence and it's the responsibility of the JudoRenderer
to ultimately convert that into ansi appropriate for the target
terminal.

* Remove now-unused ReplaceableAttributedStringBuilder and IStringBuilder

* Clean out old, now-irrelevant unit tests

* Restore window.append since we need a way to maintain scrollback

Frustrated that I forgot about this, but I've restored the test; it
doesn't pass yet, so it's ignored, but this commit also fixes some
issues with scrolling (we were not looking at the lines in the right
order).

The next thing we'll need to fix is being able to scroll past the end of
the buffer, which seems wrong; you should have at least *one* line of
output on the screen (unless of course there isn't *any* output)

* Add (currently-ignored) tests to verify prevention of over-scrolling

* Maintain scroll position on Window.append

* Finally fix scrolling

Simplifies the logic quite a bit

* Clean up compile warnings

* Add explicit kotlin-reflect dependency + bump kotlin version

* Properly clear input and window status lines when rendering

* Fix scrolling backward

* Fix connection coroutine init + dispatcher

As-is, the abstract properties didn't get initialized in time,
causing a crash.

Also, I guess since we're not using Swing we don't have a Main
dispatcher....

* Fix FlavorableCharSequence.toAttributedString and appendTo

Refs #55

* Handle disconnect more gracefully

Refs #55

* Fix terminal width resize handling

Refs #55

* Fix vertical resize handling

Refs #55

* Fix logging

Refs #55

* Import some tests and clear out the old files

Refs #55

* Clean out old JLineRenderer and related classes

Refs #55

* Fix search (again)

Refs #55

* Distinguish between "always" resize and "try" resize

Refs #55

I think the last step is to have resizing a window reactively trigger
its tabpage to re-layout!

* Handle window resize properly + prevent recursive render triggering

Refs #55

* Remove old models and replace with the new *2 versions

Refs #55

* Remove now-unused `OutputLine` class

Refs #55

* Remove last traces of jline dependency from Core

Refs #55