Rename echo -> print; includes special support for Python's `print`
`echo()` has a very specific meaning in Vim which we'd like to add
support for at some point—appending output to the buffer is all well and
good, but if you're scrolled back into the history, just writing to the
buffer won't grab your attention. This is particularly salient when
searching: we currently just write "no results" into the buffer which is
kind of useless since searching typically pulls you back into history.

It's better to change this now since nobody's using Judo but me ;)