Fix "significant" trailing whitespace getting squashed
Fixes #62

Unfortunately this reveals a slight problem with our
FlavorableCharSequence model: we don't have a way to represent a
"trailing" Flavor, IE one that doesn't live on any text, but which
should inform what style to use when rendering the rest of a line.

This should be straightforward to add, though!