Implement vertical splits (#71)
* Begin implementing HorizontalStack

Refs #64

* Implement basic horizontal stack creation/rendering

Refs #64

* Add test for multi-split rendering

Also, base the new stack's size on that of the stack it's replacing

* Add Window Commands for focusing around horizontal stacks

Refs #64

* Add normal mode mappings for new window commands

Refs #64

* Add `vsplit` scripting function

Refs #64

* Add some (ignored) tests describing some remaining issues

Refs #71

* Render *something* in the separator, to ensure garbage is cleared

* Fix cursor positioning for vertical split status line

* Add OnLayout JudoRendererEvent; fix status line rendering in vsplit

Refs #71

* Add test on resizing a vsplit

* Perform stack resize based on most-recently-resized child stack

In other words, if the user requests a window be a specific width, it
should get priority for available width allocation. We currently only do
this for HorizontalStack, but we should probably also do it for
VerticalStack

* Share window resize handling; make window resize more reliable

Less reliance on `this.width == width` checks to prevent infinite
recursion caused by a resize triggering another resize. This also should
be slightly more efficient, as well.

* Fix cursor position in multi-split layout

Includes test