Command line mode (#80)
* Begin work on a Command Line-like pseudo-mode

Refs #75

It's pretty functional! We need to render the prefix somehow, and
can probably come up with a better treatment to indicate the "Command
Line"-ness, but it's much more straightforward than I was expecting

* Add a nice prefix to rendering of Command Line mode

Refs #75

This is not the best way to implement this, actually; we should pull
prefix generation to a separate part on JudoRenderer, since currently
the prefix will get cut off in single-line input mode. Having the input
line be Flavorable is beneficial in general, though, since we could use
it for tab completion hints, for example... at some point.

* Add Command Line mode info to VimCompat.md