Fix tab completion from empty buffer after submitting command
Basically, any time we need to clear the buffer, we also need to
reset the CompletionSuggester. This is probably mostly important
for InsertMode, which we *stay in* after submitting, but just in case
I updated the other modes that use CompletionSuggester as well.

This bug has been around for a while, but I couldn't figure out how
to reproduce it until just now.