Support binding the Mapper to a specific window (#82)
* Refactor MapManager to render to a specific Window

Refs #73

Rather than defaulting to whatever window is current, it selects
the "current" window on init (eg: createEmpty) and uses that forever
unless another window is specified.

If provided a non-primary Window, that window will be cleared before
the map is rendered into it.

Furthermore, we simplify the resize mechanics to match *the specific
window* we're rendering into.

* Expose `judo.mapper.window` to scripting

Closes #73

This commit also renames the internal IJudoWindow.onSubmit to
`onSubmitFn`.  It's a little annoying, but it allows a simple
implementation of onSubmit for `IScriptWindow` that doesn't conflict
with `IJudoWindow`, so implementations of `IScriptWindow` can also
implement `IJudoWindow` for greater interop.