Handle IOExceptions more gracefully
This also simplifies the JudoConnection interface somewhat---the
refactor a while ago stopped using onError anyway, so we merged it into
onDisconnect. This also simplifies the client code a bit as well, since
there's only one handler for disconnects.

This refactor also improves unit test-ability of connection logic by
allowing forEachLine to be run as a blocking operation on the current
thread.