Support interrupting blocking ops with ctrl-c (#88)
* WIP: read keys from a Channel

* Refactor interrupt handling to be based on back-pressure

This is a bit more obvious, and simpler to reason about. In the normal
case, input is handled as soon as it's received, even a ctrl-c for
mapping purposes. However, if something is blocking the main thread and
a ctrl-c can't be delivered within some time, the onInterrupt handler is
fired so it can do its job. It may be reasonable to reduce the timeout
for this handler since it's only on ctrl-c, and because 350ms is
probably a very long time for the main thread to be blocked in normal
cases.

* Make connection cancelable with ctrl-c

Connection is now a `suspend fun` for proper structured concurrency

* WIP: restore readKey

This "works" but onMainThread is broken. I think readKey needs to
suspend....

* Add preliminary coroutine-ification for onMainThread

This introduces a couple new problems (in particular, the key reader
seems to not get stopped, and a couple tests fail) but I think is
the right direction (I even attempted it some time ago).

* Fix app not completely stopping on quit

* Clean up

* Cleanup stacktraces somewhat

* Fix assertionsWhileTyping

Last failure is in reading keys from within scripting modes (esp: the
input() fn)

* Fix deadlock executing `input()`

* Update JudoCoreJLineIntegrationTest

* Try just delaying to improve test reliability

Ultimately we may need to introduce some sort of ACK

* Remove unused postToUiQueue

* Add initial support for interrupting script execution

Useful in case you accidentally shoot yourself in the foot with
a script, so you don't have to kill the process

* Simplify error reporting when interrupting Jsr223 script execution

* Simplify

The app won't completely quit again, and I'm not entirely sure why.
It seems like the JLine NonBlockingReader is... blocking.

* Upgrade JLine

I was hoping it would fix the deadlock closing the terminal, but it does
not. Probably worth it for all the other apparent bug fixes, however.

* Ensure ScriptExecutionSandbox properly uses daemon threads

* Restore ClosedReceiveChannelException catch to fix tests

* Minor cleanup