Light Flavor implementation (#92)
* Refactor to use a Flavor factory instead of SimpleFlavor directly

If we ever do optimize the construction, this will simplify the
refactor; besides, we should've been doing this anyway if we're going to
bother with a separate Flavor interface and implementation

* Add IntFlavor implementation of Flavor

This is an inline class that can handle the common case of non-RGB
colored foreground/background coloring, etc.

* Add WrappedFlavor as a lighter-weight alternative to SimpleFlavor

Apache Spark's memory use estimator class suggests this is 24 bytes
compared to SimpleFlavor (now ExplicitFlavor)'s 48 bytes---nice! And, in
the typical case, we can just use IntFlavor, which reduces by a further
8 bytes. We will unfortunately be mostly unable to take advantage of its
inline class-ness, since we refer to Flavor by its interface, but it
still saves 8 bytes in every WrappedFlavor, since that can use the
primitive type directly.