Add preliminary work to backport linewise support to old motions
A better approach will probably be to bite the bullet and ask
IJudoWindow to render itself into a list/sequence of
FlavorableCharSequences. Since we're treating the OutputWindow as
immutable (we should prevent scrolling on append while output is
focused) we can probably cache this rendering while in the mode, and
only have to re-render if a motion scrolls.